# swagger-runtime.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Clynic API",
    "version": "v1"
  },
  "paths": {
...
    "/api/CatalogoPersonal/especialidades/clinica/{idClinica}": {
      "get": {
        "tags": [
          "CatalogoPersonal"
        ],
        "parameters": [
          {
            "name": "idClinica",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clynic.Application.DTOs.CatalogoPersonal.EspecialidadResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Citas/clinica/{idClinica}": {
      "get": {
        "tags": [
          "Citas"
        ],
        "parameters": [
          {
            "name": "idClinica",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "fechaDesde",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "fechaHasta",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "idSucursal",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "estado",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Clynic.Domain.Models.Enums.EstadoCita"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clynic.Application.DTOs.Citas.CitaResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/Clinicas": {
      "get": {
        "tags": [
          "Clinicas"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clynic.Application.DTOs.Clinicas.ClinicaResponseDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Clinicas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Clynic.Application.DTOs.Clinicas.CreateClinicaDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Clynic.Application.DTOs.Clinicas.CreateClinicaDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Clynic.Application.DTOs.Clinicas.CreateClinicaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clynic.Application.DTOs.Clinicas.ClinicaResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Clinicas/{id}": {
      "get": {
        "tags": [
          "Clinicas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clynic.Application.DTOs.Clinicas.ClinicaResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/Pacientes/clinica/{idClinica}": {
      "get": {
        "tags": [
          "Pacientes"
        ],
        "parameters": [
          {
            "name": "idClinica",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "busqueda",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clynic.Application.DTOs.Pacientes.PacienteResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/Servicios/clinica/{idClinica}": {
      "get": {
        "tags": [
          "Servicios"
        ],
        "parameters": [
          {
            "name": "idClinica",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "nombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "incluirInactivos",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clynic.Application.DTOs.Servicios.ServicioResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/Sucursales/clinica/{idClinica}": {
      "get": {
        "tags": [
          "Sucursales"
        ],
        "parameters": [
          {
            "name": "idClinica",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clynic.Application.DTOs.Sucursales.SucursalResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/Usuarios/clinica/{idClinica}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "name": "idClinica",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "nombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clynic.Application.DTOs.Usuarios.UsuarioResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Clynic.Application.DTOs.CatalogoPersonal.EspecialidadResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idClinica": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "activa": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Clynic.Application.DTOs.Citas.CitaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idClinica": {
            "type": "integer",
            "format": "int32"
          },
          "idSucursal": {
            "type": "integer",
            "format": "int32"
          },
          "idPaciente": {
            "type": "integer",
            "format": "int32"
          },
          "idEspecialidad": {
            "type": "integer",
            "format": "int32"
          },
          "nombreEspecialidad": {
            "type": "string",
            "nullable": true
          },
          "idDoctor": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nombrePaciente": {
            "type": "string",
            "nullable": true
          },
          "correoPaciente": {
            "type": "string",
            "nullable": true
          },
          "telefonoPaciente": {
            "type": "string",
            "nullable": true
          },
          "fechaHoraInicioPlan": {
            "type": "string",
            "format": "date-time"
          },
          "fechaHoraFinPlan": {
            "type": "string",
            "format": "date-time"
          },
          "duracionEstimadaMin": {
            "type": "integer",
            "format": "int32"
          },
          "fechaHoraInicioReal": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fechaHoraFinReal": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "estado": {
            "$ref": "#/components/schemas/Clynic.Domain.Models.Enums.EstadoCita"
          },
          "notas": {
            "type": "string",
            "nullable": true
          },
          "subTotal": {
            "type": "number",
            "format": "double"
          },
          "totalFinal": {
            "type": "number",
            "format": "double"
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          },
          "servicios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Clynic.Application.DTOs.Citas.CitaServicioDetalleDto"
            },
            "nullable": true
          },
          "consultaMedica": {
            "$ref": "#/components/schemas/Clynic.Application.DTOs.Citas.ConsultaMedicaResponseDto"
          }
        },
        "additionalProperties": false
      },
      "Clynic.Application.DTOs.Citas.CitaServicioDetalleDto": {
        "type": "object",
        "properties": {
          "idServicio": {
            "type": "integer",
            "format": "int32"
          },
          "nombreServicio": {
            "type": "string",
            "nullable": true
          },
          "duracionMin": {
            "type": "integer",
            "format": "int32"
          },
          "precio": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "Clynic.Application.DTOs.Citas.ConsultaMedicaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idCita": {
            "type": "integer",
            "format": "int32"
          },
          "idPaciente": {
            "type": "integer",
            "format": "int32"
          },
          "idDoctor": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "diagnostico": {
            "type": "string",
            "nullable": true
          },
          "tratamiento": {
            "type": "string",
            "nullable": true
          },
          "receta": {
            "type": "string",
            "nullable": true
          },
          "examenesSolicitados": {
            "type": "string",
            "nullable": true
          },
          "notasMedicas": {
            "type": "string",
            "nullable": true
          },
          "fechaConsulta": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "Clynic.Application.DTOs.Clinicas.ClinicaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "activa": {
            "type": "boolean"
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Clynic.Application.DTOs.Clinicas.CreateClinicaDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Clynic.Application.DTOs.Pacientes.HistorialClinicoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idPaciente": {
            "type": "integer",
            "format": "int32"
          },
          "enfermedadesPrevias": {
            "type": "string",
            "nullable": true
          },
          "medicamentosActuales": {
            "type": "string",
            "nullable": true
          },
          "alergias": {
            "type": "string",
            "nullable": true
          },
          "antecedentesFamiliares": {
            "type": "string",
            "nullable": true
          },
          "observaciones": {
            "type": "string",
            "nullable": true
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          },
          "fechaActualizacion": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Clynic.Application.DTOs.Pacientes.PacienteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idClinica": {
            "type": "integer",
            "format": "int32"
          },
          "nombres": {
            "type": "string",
            "nullable": true
          },
          "apellidos": {
            "type": "string",
            "nullable": true
          },
          "nombreCompleto": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "correo": {
            "type": "string",
            "nullable": true
          },
          "idEspecialidad": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nombreEspecialidad": {
            "type": "string",
            "nullable": true
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fechaRegistro": {
            "type": "string",
            "format": "date-time"
          },
          "historialClinico": {
            "$ref": "#/components/schemas/Clynic.Application.DTOs.Pacientes.HistorialClinicoResponseDto"
          },
          "consultasRecientes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Clynic.Application.DTOs.Citas.ConsultaMedicaResponseDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Clynic.Application.DTOs.Servicios.ServicioResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idClinica": {
            "type": "integer",
            "format": "int32"
          },
          "nombreServicio": {
            "type": "string",
            "nullable": true
          },
          "duracionMin": {
            "type": "integer",
            "format": "int32"
          },
          "precioBase": {
            "type": "number",
            "format": "double"
          },
          "activo": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Clynic.Application.DTOs.Sucursales.SucursalResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idClinica": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "activa": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Clynic.Application.DTOs.Usuarios.UsuarioResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombreCompleto": {
            "type": "string",
            "nullable": true
          },
          "correo": {
            "type": "string",
            "nullable": true
          },
          "idRol": {
            "type": "integer",
            "format": "int32"
          },
          "nombreRol": {
            "type": "string",
            "nullable": true
          },
          "descripcionRol": {
            "type": "string",
            "nullable": true
          },
          "idEspecialidad": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nombreEspecialidad": {
            "type": "string",
            "nullable": true
          },
          "activo": {
            "type": "boolean"
          },
          "debeCambiarClave": {
            "type": "boolean"
          },
          "idClinica": {
            "type": "integer",
            "format": "int32"
          },
          "nombreClinica": {
            "type": "string",
            "nullable": true
          },
          "idSucursal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nombreSucursal": {
            "type": "string",
            "nullable": true
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Clynic.Domain.Models.Enums.EstadoCita": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Ingresa el token JWT con el prefijo 'Bearer '",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}